backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pandas==2.1.4
numpy==1.25.2
python-dotenv==1.0.0
pydantic==2.5.0
httpx==0.25.2
